Requirements.txt file is invalid
Hi! I was looking at your repo and noticed that your requirements.txt file, which should enable:

```bash
$ pip install -r requirements.txt
```
is not in a valid format. Is this intentional? If it's not meant to be a python requirements file perhaps it should be named something else. I found it because I have a tool that looks for them and it choked on parsing this particular one.